refactor: :recycle: HTML fixes
Fixing up the HTML
