+++ src/main/resources/static/css/style.css
@@ -208,8 +208,13 @@ ul li a:hover{
   outline: none;
 }
 
-::placeholder{
-  color: #fff;
+#greyph::placeholder {
+  color: rgba(46, 46, 46, 0.671);
+  font-family: Arial;
+}
+
+#whiteph::placeholder {
+  color: rgba(255, 255, 255, 0.753);
   font-family: Arial;
 }
 
@@ -334,7 +339,7 @@ button:hover{
 .signup-content2 .tarikh{
   margin:10px;
   padding:10px;
-  color:#fff;
+  color:rgb(8, 8, 8);
 }
 .signup-content2  .person{
   padding:10px;
@@ -353,4 +358,65 @@ button:hover{
   width:200px;
   padding:10px;
   margin-left:200px;
+}
+.bg-white {
+  background-color: #fff !important
+}
+
+.text-black {
+  color: rgba(0, 0, 0, .5) !important
+}
+
+.text-white {
+  color: rgba(255, 255, 255, 0.945) !important
+}
+
+/* Dropdown Button */
+.dropbtn {
+  background-color: #3498DB;
+  color: white;
+  padding: 16px;
+  font-size: 16px;
+  border: none;
+  cursor: pointer;
+}
+
+/* Dropdown button on hover & focus */
+.dropbtn:hover,
+.dropbtn:focus {
+  background-color: #2980B9;
+}
+
+/* The container <div> - needed to position the dropdown content */
+.dropdown {
+  position: relative;
+  display: inline-block;
+}
+
+/* Dropdown Content (Hidden by Default) */
+.dropdown-content {
+  display: none;
+  position: absolute;
+  background-color: #f1f1f1;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
+  z-index: 1;
+}
+
+/* Links inside the dropdown */
+.dropdown-content a {
+  color: black;
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+}
+
+/* Change color of dropdown links on hover */
+.dropdown-content a:hover {
+  background-color: #ddd;
+}
+
+/* Show the dropdown menu (use JS to add this class to the .dropdown-content container when the user clicks on the dropdown button) */
+.show {
+  display: block;
 }
