+++ src/main/resources/templates/Admin.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <title layout:title-pattern="$CONTENT_TITLE - $LAYOUT_TITLE">Hibernian Rail</title>
+    <link rel="stylesheet" type="text/css" th:href="@{/css/style.css}" />
+</head> 
+<body>
+    <div class="main">
+        <div class="navbar">
+            <div class="icon">
+                <h2 class="logo">Hibernian Rail</h2>
+            </div>
+
+            <div class="menu">
+                <ul>
+                    <li><a th:href="@{/index}">HOME</a></li>
+                    <li><a href="#">ABOUT</a></li>
+                    <li><a href="#">STRIKES</a></li>
+                    <li><a th:href="@{/BookTicket}">BOOKINGS</a></li>
+                    <li><a href="#">ADMIN</a></li>
+                </ul>
+            </div>
+
+            <div class="search">
+                <input id="greyph" class="srch" type="search" name="" placeholder="Search">
+                <a href="#"> <button class="btn">Search</button></a>
+            </div>
+
+
+
+    </div>
+    <script src="https://unpkg.com/ionicons@5.4.0/dist/ionicons.js"></script>
+</body>
+</html>
