+++ src/main/resources/templates/index.html
@@ -20,12 +20,12 @@
                     <li><a href="#">ABOUT</a></li>
                     <li><a href="#">STRIKES</a></li>
                     <li><a th:href="@{/BookTicket}">BOOKINGS</a></li>
-                    <li><a href="#">ADMIN</a></li>
+                    <li><a th:href="@{/Admin}">ADMIN</a></li>
                 </ul>
             </div>
 
             <div class="search">
-                <input class="srch" type="search" name="" placeholder="Type To text">
+                <input id="greyph" class="srch" type="search" name="" placeholder="Search">
                 <a href="#"> <button class="btn">Search</button></a>
             </div>
 
@@ -37,15 +37,15 @@
                 <br> a quae totam ipsa illum minus laudantium?</p>
                 <div class="form">
                     <h2>Login Here</h2>
-                    <input type="email" name="email" placeholder="Enter Email Here">
-                    <input type="password" name="" placeholder="Enter Password Here">
+                    <input id="whiteph" type="email" name="email" placeholder="Enter Email Here">
+                    <input id="whiteph" type="password" name="" placeholder="Enter Password Here">
                     <button class="btnn"><a href="#">Login</a></button>
                     <p class="link">Don't have an account<br>
                     <a href="#">Sign up </a> here</a></p>
                 </div>
         </div>
         
-	<!-- --Plan-- -->
+	<!-- Plan -->
 	    <div class="content">
 			<h1>Payment Plan</h1><br>
 			<h2>Become a member today!</h2>
